Initial Commit
isItPrime(n)   """Tests if number passed is a prime.  isPrime() gets passed a number, and returns True if it is a prime, or False if it is not a prime"""

listOfPrimes(n)   """listOfPrimes generates, and returns a list of all the root numbers up to the number n"""

isItPrimeOptimized(n, listOfPrimes)   """isPrimeOptimized gets passed a number and a list, and only tests using the digits in the list.  Returns True if it is a prime, or False if it is not a prime"""

isItPrimeOptimized currently has performance issues